moved inactive token cleanup to ApnsServiceCache#freeUpAvailableServices
